Add git branch and branch status design
Set branch color to tango magenta with a bright white separator for all
branch statuses.
Set identical branch status to green 'v' and all non-identical statuses
to red with the existing symbol.
Preserve modified values and add them to the documentation.
Update screenshot.

This design is more in-line with sh-prompt-simple and bash-git-prompt.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>